The poster count was wrong on last poster.
This if statement was corrected so that if the posterindex is on the last poster (the number of amount of posters), it would cycle back to the first poster. Before, the webpage was showing one number higher than the amount of posters and the last poster would show two times in a row.